FIXED: #1213 transactions interfering with `wrap_predicate/4`.
More generally, transactions and snapshots should not affect some system
maintained dynamic predicates.   The wrapper used by `wrap_predicate/4`
is one example.  This also affects dynamic predicates involved in
loading sources.   This patch disconnects whether predicates are subject
to transactions from whether they are dynamic.

For now, the ability to separate the two properties is reserved to the
system.

+++ src/pl-srcfile.c
@@ -1212,7 +1212,7 @@ associateSource(SourceFile sf, Procedure proc)
 
 int
 setAttrProcedureSource(DECL_LD SourceFile sf, Procedure proc,
-		       unsigned attr, int val)
+		       uint64_t attr, int val)
 { if ( val && (attr&PROC_DEFINED) )
     associateSource(sf, proc);
